+++ gameserver/src/main/java/com/lhf/game/creature/intelligence/actionChoosers/AggroStatsChooser.java
@@ -5,8 +5,7 @@ import java.util.SortedMap;
 import java.util.TreeMap;
 
 import com.lhf.game.creature.intelligence.ActionChooser;
-import com.lhf.game.creature.intelligence.BasicAI.BattleMemories;
-import com.lhf.game.creature.intelligence.BasicAI.BattleMemories.BattleStatRecord;
+import com.lhf.game.creature.intelligence.BattleMemories;
 import com.lhf.game.enums.CreatureFaction;
 
 public class AggroStatsChooser implements ActionChooser {
